[송성학] feat: Auth guard provider 페이지들 적용
