+++ src/app/main/layout.tsx
@@ -1,6 +1,6 @@
 import AuthGuard from "@/components/AuthGuard";
 
-export default function MypageLayout({
+export default function MainLayout({
   children,
 }: {
   children: React.ReactNode;
